requirements.txt
cupy-cuda11x
numpy
